requirements.txt
# fitz
# jwt

fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
nltk
pymupdf
python-docx
numpy
pyJWT